Implemented the suggested changes.
1) moved the stepChange into common/Tour/Tour.jsx and used it directly.

2) defined a index.jsx in common/Tour.

3) Included the drop down element in the Copy Queue button during second step of tour.
