+++ simplq/src/components/common/Tour/index.jsx
@@ -1,0 +1,3 @@
+import Tour from './Tour';
+
+export default Tour;
