+++ simplq/src/components/pages/Admin/TourSteps.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
-import * as Sentry from '@sentry/react';
 
-export const getToursteps = (screenInnerWidth) => {
+const getToursteps = (screenInnerWidth) => {
   const tourProperties = {
     fontSize: '1.563rem',
     isArrowVisible: screenInnerWidth >= 1275,
@@ -63,24 +62,4 @@ export const getToursteps = (screenInnerWidth) => {
     },
   ];
 };
-
-export const stepChange = (stepNumber) => {
-  const leftArrow = document.querySelector("[data-tour-elem='left-arrow']");
-  const rightArrow = document.querySelector("[data-tour-elem='right-arrow']");
-
-    if (leftArrow && rightArrow) {
-      if (stepNumber === 0) {
-          leftArrow.childNodes[0].style.color = 'grey';
-          rightArrow.childNodes[0].style.color = 'white';
-        }
-      else if (stepNumber === 1) {
-          leftArrow.childNodes[0].style.color = 'white';
-          rightArrow.childNodes[0].style.color = 'grey';
-      }
-  }
-  else
-  {
-    Sentry.captureMessage("left-arrow or right-arrow selectors of reatTour package not found");
-  }
-
-}
+export default getToursteps;
